+++ lib/swid/backend/dart/util/produceDartTranslationUnitFromSwidClass.dart
@@ -16,7 +16,7 @@ import 'package:hydro_sdk/swid/ir/swidClass.dart';
 import 'package:hydro_sdk/swid/ir/swidType.dart';
 import 'package:hydro_sdk/swid/ir/transforms/applySuperTypes.dart';
 import 'package:hydro_sdk/swid/ir/transforms/instantiateAllGenericsAsDynamic.dart';
-import 'package:hydro_sdk/swid/ir/util/collectAllReferences.dart';
+import 'package:hydro_sdk/swid/ir/analyses/collectAllReferences.dart';
 import 'package:hydro_sdk/swid/swars/iSwarsPipeline.dart';
 
 DartTranslationUnit? produceDartTranslationUnitFromSwidClass({
