+++ lib/swid/ir/analyses/collectAllReferences.dart
