+++ lib/swid/ir/analyses/collectAllReferences.freezed.dart
