+++ test/swid/analyses/collectAllReferencesInGenericExtends_test.dart
@@ -9,7 +9,7 @@ import 'package:hydro_sdk/swid/ir/swidReferenceDeclarationKind.dart';
 import 'package:hydro_sdk/swid/ir/swidType.dart';
 import 'package:hydro_sdk/swid/ir/swidTypeFormal.dart';
 import 'package:hydro_sdk/swid/ir/swidTypeFormalValue.dart';
-import 'package:hydro_sdk/swid/ir/util/collectAllReferences.dart';
+import 'package:hydro_sdk/swid/ir/analyses/collectAllReferences.dart';
 import 'package:hydro_sdk/swid/swars/cachingPipeline.dart';
 import 'package:hydro_sdk/swid/swars/pipelineNoopCacheMgr.dart';
